Show list of available object types
Currently when the `get` command is executed without arguments it prints
the following error message:

```
Error: requires at least 1 arg(s), only received 0
```

It used to print more useful information before the introduction of the
generic implementation.

This patch improves the output, so that it will explain that the object
type is mandatory and will also list the available object types:

```shell
$ fulfillment-cli get
You must specify the type of object to get.

The following object types are available:

- clusterorders
- clusters
- clustertemplates

For example, to get the list of cluster templates:

./fulfillment-cli get clustertemplates

Use the '--help' option to get more details about the command.
```

That list of object types is obtained scanning the protobuf type
registry, so it will be automatically updated when we add new types in
the future.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>