Add `--filter` flag to the `get` command
This patch adds a new `--filter` flag to the `get` command. The filter
is a CEL expression that the server will use to decide which results to
send.

The object will be returned only if the result of the expression is
`true`.

The `this` variable in this CEL expression refers to the object. Fields
of the object can be used with this as a prefix.

For example, to find all the cluster templates that have a title
starting with `Open`:

```shell
$ ./fulfillment-cli get clustertemplates --filter "this.title.startsWith('Open')"
ID              TITLE                 DESCRIPTION
ocp_4_17_small  OpenShift 4.17 small  OpenShift 4.17 with `small` instances as worker nodes.
```

Many of the CEL operators and built-in functions are available, but not
100% of them. Documentation about that will be added to the server in
the future.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>