Add version and user agent (#25)
This patch adds a new `version` command that displays the version
identifier the tool:

```shell
$ fulfillment-cli version
0.0.4
```

It also adds the `User-Agent` header to requests. The value of the
header will be `fulfillment-service/` followed by the version number.

For releases the version number will be the result of `git describe
--tags`. For example, for release 0.0.4 the complete user agent will be
`fulfillment-cli/0.0.4`.

For unreleased versions, including versions built from source using `go
build` the version number will be the git hash.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>