Add `--template-parameter` and `--template-parameter-file`
This patch adds the following command line flags to the `create cluster`
command:

- `--template-parameter` - Sets a value for a template parameter.
- `--template-parameter-file` - Sets a value for a template parameter
  reading it from a file.

The flags can be abbreviated as `-p` for values, and `-f` for files.

The values of this flag should have the format `name=value` and
`name=file`.

The parameters are validated using the template definition and an error
message is displayed if there are issues. For example, if the user tries
to use a template that doesn't exist:

```
$ fulfillment-cli create cluster --template junk
Template 'junk' doesn't exist.

The following are some of the valid templates:

- ocp_4_17_small - OpenShift 4.17 small
- ocp_4_18_small - OpenShift 4.18 small

To see the complete list of templates use the following command:

fulfillment-cli get clustertemplates

Use the '--help' option to get more details about the command.
```

Or if the user tries to use incorrect template parameters:

```
$ ./fulfillment-cli create cluster -t ocp_4_17_small -f my_bool=junk -p hi -p =foo -p my_int=bad
There are issues with the template parameters:

- In 'hi' parameter 'hi' doesn't exist, and if it existed the value would be missing
- In '=foo' parameter name is missing
- In 'my_int=bad' value 'bad' isn't a valid 32-bit integer
- In 'my_bool=junk' file 'junk' doesn't exist
- Parameter 'my_string' is required

Valid parameters are the following:

- my_bool - boolean - My flag
- my_int - int32 - My count
- my_string - string

For more details about the template parameters run this:

./fulfillment-cli get clustertemplate ocp_4_17_small -o yaml

Use the '--help' option to get more details about the command.
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>