Add basic support for the `Host` and `HostPool` types
This patch updates the CLI to use the version of the common package that
contains the `Host` and `HostPool` types and services. As a result it
will be now possible to create, update retrieve and delete hosts and
host pools, for example:

```
$ ./fulfillment-cli get
You must specify the type of object to get.

The following object types are available:

- fulfillment.v1.Cluster
- fulfillment.v1.ClusterTemplate
- fulfillment.v1.Host
- fulfillment.v1.HostClass
- fulfillment.v1.HostPool
- fulfillment.v1.VirtualMachine
- fulfillment.v1.VirtualMachineTemplate

You can use the above fully qualified names, or the short names:

- clusters
- clustertemplates
- hostclasses
- hostpools
- hosts
- virtualmachines
- virtualmachinetemplates

For example, to get the list of clusters:

  ./fulfillment-cli get fulfillment.v1.Cluster

Or:

  ./fulfillment-cli get clusters

Note that the short names may be ambiguous if the same object type exists in different packages. In
that case the one whose fully qualified name appears first in the list will be used.

Use the '--help' option to get more details about the command.
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>