Support anonymous authentication when no token source is configured
Currently when there is no OAuth issuer advertised by the server or
expliclity specified by the user the CLI just generates an error. This
makes it difficult to use it with the development environment, where
there is no authentication configured. To avoid that this patch changes
the CLI so that in that case it will just use anonymous communication.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>