Add file parameter to the console builder
This patch adds a new `SetFile` method to the builder of the console
component, so that it will be possible to configure it to write to a
file instead of to `os.Stdout`. This is useful for unit tests, so that
the output of the console isn't mixed with the output of the testing
framework.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>