Always use template identifier with the server
A recent patch changed the `create cluster` and `create virtualmachine`
commands so that the template can be specified by identifier or by name.
But when the template is specified by name it is still necessary to pass
to the server the template identifier, not the name. That was making the
operation fail, because the server won't find a template with an
identifier matching the name. This patch fixes that issue.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>