Add lookup functionality to display tables
This patch adds a `lookup` field to the descriptions of table columns.
It indicates if the result of the CEL expression of the column is an
identifier that should be transformed into the corresponding name, if
available. For example, the clusters table contains this:

```yaml
- header: TEMPLATE
  value: this.spec.template
  type: fulfillment.v1.ClusterTemplate
  lookup: true
```

So with a template like this, which has a UUID as identifier:

```
$ fulfillment-cli get clustertemplates
ID                                    NAME           TITLE                DESCRIPTION
019abd36-b04b-75ca-b9a4-44be5a52c5f1  ocp-4-19-h100  OpenShift 4.19 H100  OpenShift 4.19 with H100 nodes
```

The output will be like this:

```
$ fulfillment-cli get clusters
ID                                    NAME        TEMPLATE       STATE        API URL  CONSOLE URL
019abd36-c5fe-77b4-b0d3-df5242633118  my-cluster  ocp-4-19-h100  PROGRESSING  -        -
```

Note how in the `TEMPLATE` column contains the name of the template, not
the UUID used as identifier.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>