Deprecate the `--cluster` option of get kubeconfig and password
This patch deprecates the `--cluster` option of the `get kubeconfig` and
`get password` commands. Instead of that the name or identifier of the
cluster will be taken from the arguments. For example:

```
$ fulfillment-cli get kubeconfig my-cluster
```

This makes them simpler to use.

We may need to revisit this, specially for the password, because in the
future we may want to get passwords for things other than clusters.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>