Support delete by name
This patch changes the `delete` command so that it will accept object
names, as long as there is no ambiguity. If there is ambiguity it will
inform the user and return an error. For example, if there are two
clusters named `my-cluster` and the user tries this:

```
$ fulfillment-cli delete cluster my-cluster
```

The result will be the following:

```
Name or identifier 'my-cluster' is ambiguous.

There are 2 matching objects:

ID                                    NAME        TEMPLATE     STATE        HUB  API URL  CONSOLE URL
019af448-ef58-7dce-b8a1-ea44c1048cda  my-cluster  my-template  PROGRESSING       -        -
019af448-f990-776b-8f40-2987dd077ec8  my-cluster  my-template  PROGRESSING       -        -

Use the identifiers instead of the names to avoid the ambiguity. For example, to delete the object
with identifier '019af448-ef58-7dce-b8a1-ea44c1048cda' use a command like this:

./fulfillment-cli delete cluster 019af448-ef58-7dce-b8a1-ea44c1048cda

Use the '--help' option to get more details about the command.
```

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>