Use device flow as the default OAuth flow for login
The device flow is more likely to succeed than the code flow because it
doesn't require the CLI and the authorization server to agree on a redirect
URI. Additionally, the device flow doesn't require the CLI to bring up an
HTTP server to receive the authorization code, which can also fail in
certain environments.

Users can still explicitly request the code flow with `--oauth-flow=code`
if desired.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>