Add `--oauth-redirect-uri` flag to login command
This change adds a new `--oauth-redirect-uri` flag to the login command,
allowing users to specify the redirect URI to use with the OAuth code flow.
The default value is `localhost:0`, which means binding to localhost on a
randomly selected port.

The `fulfillment-common` library has been updated from v0.0.23 to v0.0.24
because this new version adds the `SetRedirectUri` method to the token
source builder, which is required to implement this feature.

The redirect URI is saved to the configuration file and used when creating
the OAuth token source for subsequent commands.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>